backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asciitree==0.3.3
attrs==25.3.0
av==14.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==4.0.0
Cython==3.0.12
datasets==3.5.0
decorator==4.4.2
deepdiff==8.4.2
Deprecated==1.2.18
diffusers==0.33.1
dill==0.3.8
dm-env==1.6
dm-tree==0.1.9
dnspython==2.7.0
docker-pycreds==0.4.0
draccus==0.10.0
einops==0.8.1
email_validator==2.2.0
etils==1.12.2
Farama-Notifications==0.0.4
fastapi==0.115.12
fastapi-cli==0.0.7
fasteners==0.19
filelock==3.18.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.13.0
hf_transfer==0.1.9
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
jsonlines==4.0.0
lazy_loader==0.4
lerobot @ git+https://github.com/huggingface/lerobot.git@c10c5a0e6472986fd42ff0f9f49f29351bf3cb15
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
moviepy==1.0.3
mpmath==1.3.0
mujoco==3.3.1
mujoco-py==2.1.2.14
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.59.1
numcodecs==0.15.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orderly-set==5.4.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pfzy==0.3.4
pillow==11.2.1
proglog==0.1.11
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pygame==2.6.1
Pygments==2.19.1
pymunk==6.11.1
pynput==1.8.1
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
PyOpenGL==3.1.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rerun-sdk==0.22.1
rich==14.0.0
rich-toolkit==0.14.1
safetensors==0.5.3
scikit-image==0.22.0
scipy==1.15.2
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==75.8.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
sympy==1.13.1
tensordict==0.7.2
termcolor==2.5.0
tifffile==2025.3.30
toml==0.10.2
torch==2.6.0
TorchCodec==0.2.1
torchrl==0.7.2
torchvision==0.21.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
wandb==0.16.6
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zarr==2.18.7
zipp==3.21.0
